+++ CHANGELOG.md
@@ -5,6 +5,13 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Use a Rust struct that matches the Postgres table schema when loading a model
+  from the Postgres database.
+
 ## [0.19.0] - 2023-09-25
 
 ### Added
@@ -502,6 +509,7 @@ leading to a more streamlined system.
 
 - An initial version.
 
+[Unreleased]: https://github.com/petabi/review-database/compare/0.19.0...main
 [0.19.0]: https://github.com/petabi/review-database/compare/0.18.0...0.19.0
 [0.18.0]: https://github.com/petabi/review-database/compare/0.17.1...0.18.0
 [0.17.1]: https://github.com/petabi/review-database/compare/0.17.0...0.17.1